Renamed the createProject folder in the modules directory and updated all import paths where the folder is used.
Renamed the createProject folder in the modules directory and updated all import paths where the folder is used.
